Strip Markdown Bold or Italic

Remove markdown syntax that bold or italicizes headers and spanning headers from gtsummary tables.

Usage

strip_md_bold(x, type = "star")

strip_md_italic(x, type = "star")

Arguments

x

(gtsummary/character)
A 'gtsummary' table or character vector. When a 'gtsummary' table is passed, the column headers and spanning headers are processed.

type

(character)
Specified which type of markdown syntax to strip. Must be one or both of c("star", "underscore")

Value

character vector or gtsummary table

Examples

# Example 1 ----------------------------------
gtsummary::trial |>
  tbl_demographics(by = trt, include = age) |>
  strip_md_bold()

  

      	Drug A

N = 98
      	Drug B

N = 102
    

Age

    n
	91
	98

    Mean (SD)
	47 (15)
	47 (14)

    Median
	46
	48

    Min - Max
	6 - 78
	9 - 83

# Example 2 ----------------------------------
strip_md_bold(c("**Placebo**  \nN=45", "**High Dose**  \nN=54"))
#> [1] "Placebo  \nN=45"   "High Dose  \nN=54"