AE Rates and Counts

A mix of adverse event rates (from gtsummary::tbl_hierarchical()) and counts (from gtsummary::tbl_hierarchical_count()).

Usage

tbl_ae_rate_and_count(
  data,
  ae,
  soc,
  denominator,
  hlt = NULL,
  by = NULL,
  id = "USUBJID",
  digits = everything() ~ list(n = label_style_number(), p = label_style_number(digits =
    1, scale = 100)),
  filter = NULL,
  sort = NULL
)

# S3 method for class 'tbl_ae_rate_and_count'
add_overall(
  x,
  last = FALSE,
  col_label = "**Overall**  \nN = {style_number(N)}",
  statistic = NULL,
  digits = NULL,
  ...
)

Arguments

data

(data.frame)
a data frame.

ae

(tidy-select)
A single column name with the adverse event terms.

soc

(tidy-select)
A single column name with the system organ class.

denominator

(data.frame, integer)
used to define the denominator and enhance the output. The argument is required for tbl_hierarchical() and optional for tbl_hierarchical_count(). The denominator argument must be specified when id is used to calculate event rates.

hlt

(tidy-select)
An optional single column name with the higher level terms.

by

(tidy-select)
a single column from data. Summary statistics will be stratified by this variable. Default is NULL.

id

(tidy-select)
argument used to subset data to identify rows in data to calculate event rates in tbl_hierarchical().

digits

(formula-list-selector)
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified, default formatting is assigned via label_style_number() for the n statistic and label_style_number(digits=1, scale=100) for the p statistic.

filter, sort

Optional arguments passed to gtsummary::filter_hierarchical(filter) and gtsummary::sort_hierarchical(sort). All filtering and sorting is applied to the rates-related statistics, and not the counts.

x

(tbl_ae_rate_and_count)
a stratified 'tbl_ae_rate_and_count' table

last

(scalar logical)
Logical indicator to display overall column last in table. Default is FALSE, which will display overall column first.

col_label

(string)
String indicating the column label. Default is "**Overall** \nN = {style_number(N)}"

statistic

(formula-list-selector)
used to specify the summary statistics to display for all variables in tbl_hierarchical(). The default is everything() ~ "{n} ({p})".

...

These dots are for future extensions and must be empty.

Value

a gtsummary table

Examples

# Example 1 ----------------------------------
cards::ADAE |>
  tbl_ae_rate_and_count(
    denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
    by = TRTA,
    ae = AEDECOD,
    soc = AEBODSYS,
    # Keep rows where AEs across the row have an overall prevalence of greater than 10%
    filter = sum(n) / sum(N) > 0.10
  ) |>
  add_overall(last = TRUE)

  

Body System or Organ Class

    Dictionary-Derived Term
      	Placebo

N = 86
      	Xanomeline High Dose

N = 84
      	Xanomeline Low Dose

N = 84
      	Overall

N = 254
    

Total number of patients with at least one adverse event
	69 (80.2%)
	79 (94.0%)
	77 (91.7%)
	225 (88.6%)

Overall total number of events
	301
	455
	435
	1,191

GENERAL DISORDERS AND ADMINISTRATION SITE CONDITIONS

    Total number of patients with at least one adverse event
	21 (24.4%)
	40 (47.6%)
	47 (56.0%)
	108 (42.5%)

    Total number of events
	48
	124
	120
	292

    APPLICATION SITE ERYTHEMA
	3 (3.5%)
	15 (17.9%)
	12 (14.3%)
	30 (11.8%)

    APPLICATION SITE PRURITUS
	6 (7.0%)
	22 (26.2%)
	22 (26.2%)
	50 (19.7%)

SKIN AND SUBCUTANEOUS TISSUE DISORDERS

    Total number of patients with at least one adverse event
	21 (24.4%)
	42 (50.0%)
	42 (50.0%)
	105 (41.3%)

    Total number of events
	47
	111
	118
	276

    ERYTHEMA
	9 (10.5%)
	14 (16.7%)
	15 (17.9%)
	38 (15.0%)

    PRURITUS
	8 (9.3%)
	26 (31.0%)
	23 (27.4%)
	57 (22.4%)

    RASH
	5 (5.8%)
	11 (13.1%)
	13 (15.5%)
	29 (11.4%)