Demographics Table

This is a thin wrapper of gtsummary::tbl_summary() with the following differences:

Default summary type for continuous variables is 'continuous2'.

Number of non-missing observations is added for each variable by default and placed on the row under the header.

The tbl_summary(missing*) arguments have been renamed to tbl_demographics(nonmissing*) with updated default values.

The default footnotes from tbl_summary() are removed.

Cells with "0 (0.0%)" are converted to "0" with modify_post_fmt_fun().

Usage

tbl_demographics(
  data,
  by = NULL,
  label = NULL,
  statistic = list(all_continuous() ~ c("{mean} ({sd})", "{median}", "{min} - {max}"),
    all_categorical() ~ "{n} ({p}%)"),
  digits = NULL,
  type = NULL,
  value = NULL,
  nonmissing = c("always", "ifany", "no"),
  nonmissing_text = "n",
  nonmissing_stat = "{N_nonmiss}",
  sort = all_categorical(FALSE) ~ "alphanumeric",
  percent = c("column", "row", "cell"),
  include = everything()
)

Arguments

data

(data.frame)
A data frame.

by

(tidy-select)
A single column from data. Summary statistics will be stratified by this variable. Default is NULL.

label

(formula-list-selector)
Used to override default labels in summary table, e.g. list(age = "Age, years"). The default for each variable is the column label attribute, attr(., 'label'). If no label has been set, the column name is used.

statistic

(formula-list-selector)
Specifies summary statistics to display for each variable. The default is list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}%)"). See below for details.

digits

(formula-list-selector)
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified, default formatting is assigned via assign_summary_digits(). See below for details.

type

(formula-list-selector)
Specifies the summary type. Accepted value are c("continuous", "continuous2", "categorical", "dichotomous"). If not specified, default type is assigned via assign_summary_type(). See below for details.

value

(formula-list-selector)
Specifies the level of a variable to display on a single row. The gtsummary type selectors, e.g. all_dichotomous(), cannot be used with this argument. Default is NULL. See below for details.

nonmissing, nonmissing_text, nonmissing_stat

Arguments dictating how and if missing values are presented:

nonmissing: must be one of c("always", "ifany", "no")

nonmissing_text: string indicating text shown on non-missing row. Default is "n"

nonmissing_stat: statistic to show on non-missing row. Default is "{N_nonmiss}". Possible values are N_nonmiss, N_miss, N_obs,, p_nonmiss p_miss.

sort

(formula-list-selector)
Specifies sorting to perform for categorical variables. Values must be one of c("alphanumeric", "frequency"). Default is all_categorical(FALSE) ~ "alphanumeric".

percent

(string)
Indicates the type of percentage to return. Must be one of c("column", "row", "cell"). Default is "column".

include

(tidy-select)
Variables to include in the summary table. Default is everything().

Value

a 'gtsummary' table

Examples

# Example 1 ----------------------------------
trial |>
  tbl_demographics(
    by = trt,
    include = c(age, grade)
  ) |>
  add_overall()

  

	
      	Overall

N = 200
      	Drug A

N = 98
      	Drug B

N = 102
    

	Age

	    n
	189
	91
	98

	    Mean (SD)
	47 (14)
	47 (15)
	47 (14)

	    Median
	47
	46
	48

	    Min - Max
	6 - 83
	6 - 78
	9 - 83

	Grade

	    n
	200
	98
	102

	    I
	68 (34%)
	35 (36%)
	33 (32%)

	    II
	68 (34%)
	32 (33%)
	36 (35%)

	    III
	64 (32%)
	31 (32%)
	33 (32%)