Roche Theme

A gtsummary theme for Roche tables

Calls the gtsummary::theme_gtsummary_compact() theme.

Uses label_roche_pvalue() as the default formatting function for all p-values.

Defaults to a mono-spaced font for gt tables.

Font size defaults are 8 points for all the table by the footers that are 7 points.

Border defaults to flextable::fp_border_default(width = 0.5).

Usage

theme_gtsummary_roche(set_theme = TRUE)

Arguments

set_theme

(scalar logical)
Logical indicating whether to set the theme. Default is TRUE. When FALSE the named list of theme elements is returned invisibly

Value

theme list

Examples

theme_gtsummary_roche()
#> Setting theme "Roche"

tbl_demographics(trial, by = trt, include = c(age, grade))

  

      	Drug A

N = 98
      	Drug B

N = 102
    

Age

    n
	91
	98

    Mean (SD)
	47 (15)
	47 (14)

    Median
	46
	48

    Min - Max
	6 - 78
	9 - 83

Grade

    n
	98
	102

    I
	35 (35.7%)
	33 (32.4%)

    II
	32 (32.7%)
	36 (35.3%)

    III
	31 (31.6%)
	33 (32.4%)

reset_gtsummary_theme()